refactor(base-app): extract reading tools into a separate entity
