--- apps/base-app/src/constants/defaultSettingsReadingTool.ts
@@ -1,10 +1,0 @@
-import type { SettingsReadingTool } from '@/interfaces/settingsReadingTool'
-
-const buildDefaultSettingsReadingTool = (): SettingsReadingTool => {
-  return {
-    opacity: '33',
-    thickness: '1'
-  }
-}
-
-export { buildDefaultSettingsReadingTool }
