+++ apps/base-app/src/entities/readingTools/index.ts
@@ -1,0 +1,2 @@
+export type { SettingsReadingTool } from './model/settingsReadingTool'
+export { buildDefaultSettingsReadingTool } from './model/defaultSettingsReadingTool'
