+++ apps/base-app/src/entities/readingTools/model/defaultSettingsReadingTool.ts
@@ -1,0 +1,8 @@
+import type { SettingsReadingTool } from './settingsReadingTool'
+
+export const buildDefaultSettingsReadingTool = (): SettingsReadingTool => {
+  return {
+    opacity: '33',
+    thickness: '1'
+  }
+}
