+++ apps/base-app/src/entities/readingTools/model/settingsReadingTool.ts
